requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
arrow==1.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
backports.ssl==0.0.9
backports.ssl_match_hostname==3.7.0.1
bcrypt==4.1.2
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
cryptography==41.0.7
dict2xml==1.7.5
dill==0.3.8
distro==1.9.0
f5-icontrol-rest==1.3.13
frozenlist==1.4.1
future==0.18.3
genie==24.1
genie.libs.clean==24.1
genie.libs.conf==24.1
genie.libs.filetransferutils==24.1
genie.libs.health==24.1
genie.libs.ops==24.1
genie.libs.parser==24.1
genie.libs.robot==24.1
genie.libs.sdk==24.1
genie.telemetry==24.1
genie.trafficgen==24.1
gitdb==4.0.11
GitPython==3.1.41
grpcio==1.60.1
idna==3.6
importlib-metadata==7.0.1
IxNetwork==9.30.2212.8
ixnetwork-restpy==1.1.12
Jinja2==3.1.3
jsonpickle==3.0.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
junit-xml==1.9
junos-eznc==2.7.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.5
napalm==4.1.0
ncclient==0.6.15
netaddr==0.10.1
netmiko==4.3.0
netutils==1.6.0
ntc_templates==4.1.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
paramiko==3.4.0
pathspec==0.12.1
pillow==10.2.0
ply==3.11
prettytable==3.9.0
protobuf==4.25.2
psutil==5.9.8
pyarrow==14.0.2
pyasn1==0.4.8
pyats==24.1
pyats.aereport==24.1
pyats.aetest==24.1
pyats.async==24.1
pyats.connections==24.1
pyats.contrib==24.1
pyats.datastructures==24.1
pyats.easypy==24.1
pyats.kleenex==24.1
pyats.log==24.1
pyats.reporter==24.1
pyats.results==24.1
pyats.robot==24.1
pyats.tcl==24.1
pyats.topology==24.1
pyats.utils==24.1
pycparser==2.21
pydeck==0.8.1b0
pyeapi==1.0.2
pyftpdlib==1.5.9
Pygments==2.17.2
PyNaCl==1.5.0
pynetbox==7.3.3
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyserial==3.5
pysmi-lextudio==1.1.13
pysnmp-lextudio==5.0.29
pysnmpcrypto==0.0.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==3.14.2
python-slugify==8.0.4
python-socketio==4.6.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
rest.connector==24.1
rich==13.7.0
robotframework==7.0
rpds-py==0.17.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scp==0.14.5
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
tenacity==8.2.3
text-unidecode==1.3
textfsm==1.1.3
tftpy==0.8.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.2
transitions==0.9.0
ttp==0.9.5
ttp-templates==0.3.6
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
unicon==24.1
unicon.plugins==24.1
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
websocket-client==1.7.0
xlrd==1.2.0
XlsxWriter==3.1.9
xlwt==1.3.0
xmltodict==0.13.0
yamllint==1.34.0
yamlordereddictloader==0.4.2
yang.connector==24.1
yarl==1.9.4
zipp==3.17.0
